Support transitive availability/visibility chains.
This corresponds to changes added to the Vulkan memory model appendix
released in the Vulkan 1.1.98 spec update. The chains are an optional
feature added to v3 of VK_KHR_vulkan_memory_model.

Transitive av/vis chains allow, for example, a single representative
invocation to make a whole workgroup's writes available outside of the
workgroup.