Loosen consistent[X] predicate
Make the "consistent" predicate accept reads that are not in location
order, but still forbid reads from locations that are shadowed by another
write.

Also simplify tests to have a simpler and more principled interpretation.
A typical failing data race test has a satisfiable solution that is
consistent and has a data race, and no solution that is consistent and
race-free. A typical pass test simply has the predicates reversed: there
is a consistent data-race free solution, and no consistent solution with
a data race.

This version of the PR is incomplete, as there are three tests that need
further attention: mp3acqrel, releaseseq3, and releaseseq4. They all pass
but could probably be refined.