Add remaining tests
The mp, fencefence2, ssw0, and ssw1 tests were just an oversight, the other three had a more complex predicate. On examination, I don't think the is_acyclic is essential to what's being tested, and removing that conjuct form a NOSOLUTION is a stricter test, so given that they pass, I'm satisfied.