Align avvis semantics with the specification
This commit updates the definition of avvis semantics to align
with the Vulkan specification. The bidirectional avvisinc relation
is now replaced with a unidirectional one. For availability
semantics, the avvisinc relation is generated for operations
ordered before the current operation but not for the operation
itself. For visibility semantics, the definition is symmetric.

This commit changes the behavior of test mpinscope3, mpnotinscope3,
and mpnotinscope6 from safe to unsafe under the memory model.

Signed-off-by: Natalia Gavrilenko <natalia.gavrilenko@huawei.com>